requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
openai==1.34.0
chromadb==0.5.0
python-dotenv==1.0.0
colorama==0.4.6
tiktoken==0.7.0
langchain-huggingface==0.2.5
sentence-transformers==5.1.2
langchain-ollama==1.0.1
numpy<2.0.0
huggingface-hub<1.0.0
fastapi==0.111.0
uvicorn==0.30.1
langchain-groq==0.1.3


